Add files via upload
the uploaded zip file conduct the basic MVC frames for the ThinkPhp opensource works. The TP work comes out with Home MVC frames automatically when set up successfully,  Education was considered by the Home frame. With the responsive web matched to this TP, the entity web codes resource would be launched successfully. welcome to the page for commits.